+++ src/database/membershipTags.ts
@@ -10,6 +10,7 @@ function mapResultsToTags(result: any) : MembershipTag[] {
     result?.forEach((row : any) => {
         const tag : MembershipTag = {
             id: row.membership_tag_id,
+            tenantId: row.tenant_id,
             value: row.membership_tag,
             membershipId: row.membership_id,
             count: row.tag_count,
@@ -19,14 +20,14 @@ function mapResultsToTags(result: any) : MembershipTag[] {
     return tags;
 }
 
-export async function getMembershipTags(membershipId: number) : Promise<MembershipTag[]> {
+export async function getMembershipTags(membershipId: number, tenantId: string) : Promise<MembershipTag[]> {
     let result;
 
     try {
         [result] = await getPool().query<RowDataPacket[]>(
             // eslint-disable-next-line max-len
-            'select membership_tag_id, membership_tag, membership_id from membership_tags where membership_id = ? order by membership_tag',
-            [membershipId],
+            'select membership_tag_id, membership_tag, membership_id from membership_tags where membership_id = ? and tenant_id = ? order by membership_tag',
+            [membershipId, tenantId],
         );
     } catch (e) {
         logger.error(`DB error getting tags for membership ID ${membershipId}`, e);
@@ -34,49 +35,54 @@ export async function getMembershipTags(membershipId: number) : Promise<Membersh
     return mapResultsToTags(result);
 }
 
-export async function createMembershipTag(membershipId: number, tags: string[]) : Promise<MembershipTag[]> {
+export async function createMembershipTag(
+    membershipId: number, tags: string[],
+    tenantId: string,
+) : Promise<MembershipTag[]> {
     try {
-        const insertSql = 'insert into membership_tags(membership_id, membership_tag) values (?, ?)';
+        const insertSql = 'insert into membership_tags(membership_id, membership_tag, tenant_id) values (?, ?, ?)';
         tags.forEach(async (tag) => {
-            await getPool().query<OkPacket>(insertSql, [membershipId, tag]);
+            await getPool().query<OkPacket>(insertSql, [membershipId, tag, tenantId]);
         });
     } catch (error) {
         logger.error(`Error inserting tags ${tags.join(',')} for membershipID ${membershipId}`, error);
     }
-    const newTags = await getMembershipTags(membershipId);
+    const newTags = await getMembershipTags(membershipId, tenantId);
     return newTags;
 }
 
-export async function deleteMembershipTag(membershipId: number, tags: string[]) : Promise<MembershipTag[]> {
+// eslint-disable-next-line max-len
+export async function deleteMembershipTag(membershipId: number, tags: string[], tenantId: string) : Promise<MembershipTag[]> {
     try {
-        const deleteSql = 'delete from membership_tags where membership_id = ? and membership_tag = ?';
+        // eslint-disable-next-line max-len
+        const deleteSql = 'delete from membership_tags where membership_id = ? and membership_tag = ? and tenant_id = ?';
         tags.forEach(async (tag) => {
-            await getPool().query<OkPacket>(deleteSql, [membershipId, tag]);
+            await getPool().query<OkPacket>(deleteSql, [membershipId, tag, tenantId]);
         });
     } catch (error) {
         logger.error(`Error deleting tags ${tags.join(',')} for membershipID ${membershipId}`, error);
     }
-    const newTags = await getMembershipTags(membershipId);
+    const newTags = await getMembershipTags(membershipId, tenantId);
     return newTags;
 }
 
-export async function cleanMembershipTags(membershipId: number) : Promise<MembershipTag[]> {
+export async function cleanMembershipTags(membershipId: number, tenantId: string) : Promise<MembershipTag[]> {
     try {
-        const deleteSql = 'delete from membership_tags where membership_id = ?';
-        await getPool().query<OkPacket>(deleteSql, [membershipId]);
+        const deleteSql = 'delete from membership_tags where membership_id = ? and tenant_id = ?';
+        await getPool().query<OkPacket>(deleteSql, [membershipId, tenantId]);
     } catch (error) {
         logger.error(`Error deleting tags for membershipID ${membershipId}`, error);
     }
-    const newTags = await getMembershipTags(membershipId);
+    const newTags = await getMembershipTags(membershipId, tenantId);
     return newTags;
 }
 
-export async function getUniqueTags() : Promise<MembershipTag[]> {
+export async function getUniqueTags(tenantId: string) : Promise<MembershipTag[]> {
     let result;
 
     try {
         // eslint-disable-next-line max-len
-        [result] = await getPool().query<RowDataPacket[]>('select distinct membership_tag, count(*) tag_count from membership_tags group by membership_tag');
+        [result] = await getPool().query<RowDataPacket[]>('select distinct membership_tag, count(*) tag_count from membership_tags where tenant_id = ? group by membership_tag', [tenantId]);
     } catch (e) {
         logger.error('DB error getting unique membership tags', e);
     }
